Banking_System/requirements.txt
asgiref==3.8.1
Django==5.0
djangorestframework==3.15.1
pillow==10.3.0   # Pillow library for image and File fields
psycopg2==2.9.9
sqlparse==0.5.0
tzdata==2024.1
